src/requirements.txt
httplib2==0.19.0
requests==2.32.0
gcloud==0.18.3
google-cloud-storage==2.13.0
google-cloud-bigquery==3.24.0
pyarrow==16.1.0
db-dtypes==1.2.0
uvicorn==0.24.0.post1
fastapi==0.109.1
openpyxl==3.1.2
retry==0.9.2
pandas==2.2.2
colorlog==6.8.2
python-dotenv==1.0.1
pytest==8.2.2
pyfiglet==0.8.post1
setuptools>=65.0.0
termcolor==2.4.0
